% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munisbr_package.R
\docType{package}
\name{munisbr}
\alias{munisbr}
\alias{_PACKAGE}
\alias{munisbr-package}
\title{munisbr: Download Official Socioeconomic Data of Brazilian Municipalities}
\description{
Downloads  official data of socioeconomic and population characteristics of
Brazilian municipalities. The package offers a set of convenient functions to
download IBGE's data through the "sidrar package.
}
\section{Usage}{

Please check the vignettes and data documentation on the
\href{https://ipeagit.github.io/munisbr/index.html}{website}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ipeaGIT/munisbr}
  \item \url{https://ipeagit.github.io/munisbr/}
  \item Report bugs at \url{https://github.com/ipeaGIT/munisbr/issues}
}

}
\author{
\strong{Maintainer}: Rafael H. M. Pereira \email{rafa.pereira.br@gmail.com} (\href{https://orcid.org/0000-0003-2125-7465}{ORCID})

}
\keyword{internal}
