% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gdp.R
\name{read_gdp}
\alias{read_gdp}
\title{GDP data}
\usage{
read_gdp(year)
}
\arguments{
\item{year}{Numeric. Year in the formart YYYY. It can also take a vector with
the period of the data, such as \code{c(2010, 2015)}. Defaults to the last
year available.}
}
\description{
Download annual GDP estimates of Brazilian municipalities.
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(munisbr)

gdp_br <- read_gdp(year = 2020)
head(gdp_br)
\dontshow{\}) # examplesIf}
}
\concept{Economic data}
