% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_population.R
\name{read_population}
\alias{read_population}
\title{Population data}
\usage{
read_population(year)
}
\arguments{
\item{year}{Numeric. Year in the formart YYYY. It can also take a vector with
the period of the data, such as \code{c(2010, 2015)}. Defaults to the last
year available.}
}
\description{
Download annual population estimates of Brazilian municipalities.
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(munisbr)

pop_br <- read_population(year = 2020)
head(pop_br)
\dontshow{\}) # examplesIf}
}
\concept{Demographic data}
